ai-services/rag/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
qdrant-client>=1.12.0
sentence-transformers==3.0.0
langchain==0.2.0
langchain-text-splitters==0.2.0
pypdf==4.2.0
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0